backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
Jinja2==3.1.4
langchain==0.3.7
langchain-core==0.3.15
httpx==0.27.2
python-docx==1.1.2
reportlab==4.2.5
starlette==0.41.2
itsdangerous==2.2.0
orjson==3.10.7
